Add files via upload
there are two types of data visualization:

Exploratory data visualization: we create graphs for ourselves to better understand and

explore data.

Explanatory data visualization: we create graphs for others to inform, make a point, or tell a story.

Throughout the Project, we focused on explanatory data visualization and learned the following:

How to use information design principles (familiarity and maximizing the data-ink ratio) to create better graphs for an audience.

About the elements of a story and how to create storytelling data visualizations using

Matplotlib. How to guide the audience's attention with pre-attentive attributes.

How to use Matplotlib built-in styles - with a case study on the Five ThirtyEight style.